requirements.txt
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cryptography==46.0.5
idna==3.11
numpy==2.4.3
opencv-python==4.13.0.92
packaging==26.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.1
pycparser==3.0
PyMuPDF==1.27.2.2
pypdfium2==5.6.0
pytesseract==0.3.13
requests==2.32.5
urllib3==2.6.3
